refactor: move truePBR to it's own feature
