+++ src/Feature.cpp
@@ -13,6 +13,7 @@
 #include "Features/SubsurfaceScattering.h"
 #include "Features/TerrainBlending.h"
 #include "Features/TerrainOcclusion.h"
+#include "Features/TruePBR.h"
 #include "Features/WaterCaustics.h"
 #include "Features/WaterParallax.h"
 #include "Features/WetnessEffects.h"
@@ -126,7 +127,8 @@ const std::vector<Feature*>& Feature::GetFeatureList()
 		TerrainOcclusion::GetSingleton(),
 		ScreenSpaceGI::GetSingleton(),
 		Skylighting::GetSingleton(),
-		TerrainBlending::GetSingleton()
+		TerrainBlending::GetSingleton(),
+		TruePBR::GetSingleton()
 	};
 
 	static std::vector<Feature*> featuresVR(features);
