+++ src/FeatureBuffer.cpp
@@ -6,6 +6,7 @@
 #include "Features/LightLimitFix.h"
 #include "Features/Skylighting.h"
 #include "Features/TerrainOcclusion.h"
+#include "Features/TruePBR.h"
 #include "Features/WetnessEffects.h"
 
 template <class... Ts>
@@ -34,5 +35,5 @@ std::pair<unsigned char*, size_t> GetFeatureBufferData()
 		WetnessEffects::GetSingleton()->GetCommonBufferData(),
 		LightLimitFix::GetSingleton()->GetCommonBufferData(),
 		Skylighting::GetSingleton()->cbData,
-		State::GetSingleton()->pbrSettings);
+		TruePBR::GetSingleton()->settings);
 }
