+++ src/Hooks.h
@@ -2,5 +2,6 @@
 
 namespace Hooks
 {
+	bool hk_BSShader_BeginTechnique(RE::BSShader* shader, uint32_t vertexDescriptor, uint32_t pixelDescriptor, bool skipPixelShader);
 	void Install();
 }
